// apps/web/src/types/meilisearch.ts
// Generated by https://quicktype.io

export interface PodcastSearch {
  id: string;
  title: string;
  description: string;
  createdAt: string;
  updatedAt: string;
  audio: Audio;
  cover: Audio;
  user: User;
}

export interface Audio {
  id: number;
  name: string;
  alternativeText: string;
  caption: string;
  width: number | null;
  height: number | null;
  formats: Formats | null;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl: null | string;
  provider: string;
  provider_metadata: ProviderMetadata;
  folderPath: string;
  createdAt: string;
  updatedAt: string;
}

export interface Formats {
  large: Large;
  small: Large;
  medium: Large;
  thumbnail: Large;
}

export interface Large {
  ext: string;
  url: string;
  hash: string;
  mime: string;
  name: string;
  path: null;
  size: number;
  width: number;
  height: number;
  provider_metadata: ProviderMetadata;
}

export interface ProviderMetadata {
  public_id: string;
  resource_type: ResourceType;
}

export enum ResourceType {
  Image = "image",
  Video = "video",
}

export interface User {
  id: number;
  username: string;
  email: string;
  provider: string;
  password: string;
  resetPasswordToken: null;
  confirmationToken: null;
  confirmed: boolean;
  blocked: boolean;
  createdAt: string;
  updatedAt: string;
}